@@ -1,6 +1,6 @@
 import { type FC } from "react";
 import NextImage, { type ImageProps as NextImageProps } from "next/image";
-import { shimmer, toBase64 } from "@chia/utils/shimmer.util";
+// import { shimmer, toBase64 } from "@chia/utils/shimmer.util";
 
 const Image: FC<NextImageProps> = (props) => {
   const { alt, ...rest } = props;
@@ -8,8 +8,8 @@ const Image: FC<NextImageProps> = (props) => {
     <NextImage
       alt={alt}
       aria-label={alt}
-      blurDataURL={`data:image/svg+xml;base64,${toBase64(shimmer(700, 700))}`}
-      placeholder="blur"
+      // blurDataURL={`data:image/svg+xml;base64,${toBase64(shimmer(700, 700))}`}
+      // placeholder="blur"
       {...rest}
     />
   );
