@@ -2,7 +2,6 @@ import path from "path";
 import type { PostFrontMatter, PostSource } from "@chia/shared/types";
 import { POSTS_PATH } from "@chia/shared/constants";
 import { serialize } from "next-mdx-remote/serialize";
-import { minify } from "uglify-js";
 import pMap from "p-map";
 import glob from "fast-glob";
 import { getPostData } from "../repositories";
@@ -50,23 +49,10 @@ export const getPost = async (slug: string): Promise<PostSource> => {
     },
   });
 
-  // HACK: next-mdx-remote v4 doesn't (yet?) minify compiled JSX output, see:
-  // https://github.com/hashicorp/next-mdx-remote/pull/211#issuecomment-1013658514
-  // ...so for now, let's do it manually (and conservatively) with uglify-js when building for production.
-  const compiledSource =
-    process.env.NODE_ENV === "production"
-      ? minify(source.compiledSource, {
-          toplevel: true,
-          parse: {
-            bare_returns: true,
-          },
-        }).code
-      : source.compiledSource;
-
   return {
     frontMatter,
     source: {
-      compiledSource,
+      compiledSource: source.compiledSource,
     },
   };
 };
